Move Event date formatting methods to Formatable module
